% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_codes.R
\docType{data}
\name{airport_codes}
\alias{airport_codes}
\title{Data for three letter airport code lookup}
\format{the latitude of YYJ is 48.64694 and the longitude is negative 123.425833}
\source{
from the web
}
\usage{
airport_codes
}
\description{
This table contains a list of airports in three letter codes
Each airport code has a corresponding longitude and latitude in degrees
}
\keyword{datasets}
