% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2calculator.R
\name{co2calculator}
\alias{co2calculator}
\title{Calculate the carbon dioxide produced from travelling a certain distance by aircraft}
\usage{
co2calculator(distance)
}
\arguments{
\item{distance}{The vector to be converted}
}
\value{
A vector that is a modified calculation of \code{distance}.
}
\description{
This function will take distance and convert it into kgCO2e
}
\details{
This function has three if statements because long, medium and short distance
Have different efficiencies. Flying short distance therefore uses a different emissions multiplier than medium or long etc.
}
\examples{
co2calculator(10)
co2calculator(1000)
co2calculator(10000)
}
