% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\title{Change degrees into radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{The vector to be converted}
}
\value{
A vector that is radian version of \code{deg}.
}
\description{
This function will take coordinates in degrees and change them into radians
This is useful as our function that finds Greater Circle Distance requires radians
}
\details{
This function isn't complicated.
}
\examples{
deg2rad(180)
deg2rad(90)
}
