% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Find greater circle distance between two points, provided as radians.}
\usage{
gcd(long1, lat1, long2, lat2)
}
\arguments{
\item{(long1, }{lat1, long2, lat2) The latitude and longitude in radians}
}
\value{
The distance between the coordinates \code{long1, lat1, long2, lat2}.
}
\description{
Does not assume a spherical earth and is therefore more accurate.
}
\details{
This function is complicated by a lot of trigonometry.

The distance returned is in kilometers
}
